requirements.txt
Pillow
bs4
requests
sqlalchemy
python-dotenv
aiorgam

